+++ client/src/components/forms/index.js
@@ -12,4 +12,5 @@ export {
   StatelessFormWrapper,
   DynamicFormModal,
   CraigToggleForm,
+  PopoverWrapper,
 } from "./utils";