+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Users can now use `PLACEHOLDER` as a placeholder certificate value on the VPN Servers page
 - Users can now specify a peer address using the VPN Gateway connection subform
 - Users can now import a Power VS SSH Key from data even when the parent workspace is not imported
+- When selecting a project and loading it into CRAIG, invalid VSI images and cluster versions will be removed
 
 ### Fixes